% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustA.R
\name{adjustA}
\alias{adjustA}
\title{Adjusts alpha vector for contamination karyotypes}
\usage{
adjustA(a, c1, c2, g, correction = 1e-06)
}
\arguments{
\item{a}{vector to change}

\item{c1}{X multiplier for karyotype 1}

\item{c2}{X multiplier for karyotype 2}

\item{g}{mixing parameter}

\item{correction}{error rate for Y mapping for non-Y karyotypes}
}
\value{
adjusted alpha vector
}
\description{
Adjusts alpha vector for contamination karyotypes
}
