% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound.R
\name{bound}
\alias{bound}
\title{bounds number to be in [0,1] (contamination estimation)}
\usage{
bound(x)
}
\arguments{
\item{x}{number}
}
\value{
bounded number
}
\description{
bounds number to be in [0,1] (contamination estimation)
}
