% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callKaryotypes.R
\name{callKaryotypes}
\alias{callKaryotypes}
\title{Take prior and observed counts and calls karyotypes and Z-scores}
\usage{
callKaryotypes(indat, inDirichlet, p_contamination = 0.1)
}
\arguments{
\item{indat}{full input tibble}

\item{inDirichlet}{associated Dirichlet parameters}

\item{p_contamination}{the assumed probability of contamination}
}
\value{
karyotype calls
}
\description{
Take prior and observed counts and calls karyotypes and Z-scores
}
