% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineData.R
\name{combineData}
\alias{combineData}
\title{combine data}
\usage{
combineData(calls.auto, calls.sca, z.scores, printMissingIDs = FALSE)
}
\arguments{
\item{calls.auto}{a list or predefined piped string of sample IDs to look for}

\item{calls.sca}{the autosomal karyotype calls for the samples}

\item{z.scores}{the sex chromosomal karyotype calls for the samples}

\item{printMissingIDs}{print the IDs of individuals not in all 3 input files (or just numbers)?}
}
\value{
combined data
}
\description{
Combine both autosomal and sex chromosomal analyses into one output
}
