% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddirichletultinomial.R
\name{ddirichletultinomial}
\alias{ddirichletultinomial}
\title{calculate Dirichlet probabilities with filtering}
\usage{
ddirichletultinomial(nvector, avector, cvector, log = T, correction = 1e-06)
}
\arguments{
\item{nvector}{observed read counts}

\item{avector}{alpha vector to change}

\item{cvector}{change vector for new karyotype}

\item{log}{return value in log space}

\item{correction}{error rate for Y mapping for non-Y karyotypes}
}
\value{
Dirichlet probabilities
}
\description{
calculate Dirichlet probabilities with filtering
}
