% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.mle_filter.R
\name{dirichlet.mle_filter}
\alias{dirichlet.mle_filter}
\title{calculate Dirichlet parameters with filtering}
\usage{
dirichlet.mle_filter(y, whichK)
}
\arguments{
\item{y}{input values}

\item{whichK}{which parameter to return}
}
\value{
parameters
}
\description{
calculate Dirichlet parameters with filtering
}
