% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGamma.R
\name{estimateGamma}
\alias{estimateGamma}
\title{Estimate contamination from XX and XY based on observed counts}
\usage{
estimateGamma(nvector, avector, c1, c2, correction)
}
\arguments{
\item{nvector}{observed read counts}

\item{avector}{vector to change}

\item{c1}{X multiplier for karyotype 1}

\item{c2}{X multiplier for karyotype 2}

\item{correction}{error rate for Y mapping for non-Y karyotypes}
}
\value{
estimated contamination
}
\description{
Estimate contamination from XX and XY based on observed counts
}
