% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{example_data}
\format{
A data frame with 266 rows and 26 variables:
}
\source{
"Ben Rohrlach"
}
\usage{
example_data
}
\description{
An example data set for RChASM
}
\keyword{datasets}
