% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterOutliers.R
\name{filterOutliers}
\alias{filterOutliers}
\title{filter for outliers}
\usage{
filterOutliers(v)
}
\arguments{
\item{v}{vector to filter}
}
\value{
vector for filtering
}
\description{
filter for outliers
}
