% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeC.R
\name{makeC}
\alias{makeC}
\title{make c vector for new karyotypes}
\usage{
makeC(x, y, n)
}
\arguments{
\item{x}{value to replace 1}

\item{y}{location to make replacement}

\item{n}{length of vector}
}
\value{
c vector
}
\description{
make c vector for new karyotypes
}
