% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDirichlet.R
\name{makeDirichlet}
\alias{makeDirichlet}
\title{function to make a Dirichlet prior}
\usage{
makeDirichlet(indat, refType, min_reads = 30000, max_reads = 1e+09)
}
\arguments{
\item{indat}{full input tibble}

\item{refType}{"auto"="autosomal"; "sca"="sex chromosomal"; "diagnostic"}

\item{min_reads}{min number of reads per sample}

\item{max_reads}{max number of reads per sample}
}
\value{
Dirichlet prior
}
\description{
function to make a Dirichlet prior
}
