% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeZscores.R
\name{makeZscores}
\alias{makeZscores}
\title{Calculates Z-scores per chromosome}
\usage{
makeZscores(indat, refType = "auto", min_reads = 30000, max_reads = 1e+09)
}
\arguments{
\item{indat}{full input tibble}

\item{refType}{"auto"="autosomal"; "sca"="sex chromosomal"; "diagnostic"}

\item{min_reads}{min number of reads per sample}

\item{max_reads}{max number of reads per sample}
}
\value{
Z-scores
}
\description{
Calculates Z-scores per chromosome
}
