% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printChASM.R
\name{printChASM}
\alias{printChASM}
\title{A function to print the result of the combined analysis to screen}
\usage{
printChASM(inChASM, lines = 20)
}
\arguments{
\item{inChASM}{the result of a full ChASM analysis (output from runChASM)}

\item{lines}{the number of lines to print to screen}
}
\value{
print output
}
\description{
A function to print the result of the combined analysis to screen
}
\examples{
example_calls <- runChASM(rawReadCountsIn = example_data)
printChASM(inChASM = example_calls, lines = 10)

}
