% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processReadCounts.R
\name{processReadCounts}
\alias{processReadCounts}
\title{processReadCounts}
\usage{
processReadCounts(rawReadCountsIn, refType, minTotal = 1000, minProtocol = 30)
}
\arguments{
\item{rawReadCountsIn}{the reads counts via KP}

\item{refType}{for capture, use only on-target (F)?}

\item{minTotal}{minimum number of protocol counts to be included}

\item{minProtocol}{"auto"="autosomal"; "sca"="sex chromosomal"}
}
\value{
outreadcounts
}
\description{
Task: takes read counts and makes useful file
}
