% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runChASM.R
\name{runChASM}
\alias{runChASM}
\title{Title}
\usage{
runChASM(
  rawReadCountsIn,
  minSamplesPerProtocol = 30,
  min_reads = 60000,
  max_reads = 1e+09,
  p_contamination = 0.01,
  show_plot = TRUE,
  printMissingIDs = FALSE
)
}
\arguments{
\item{rawReadCountsIn}{the reads counts for each chromosome}

\item{minSamplesPerProtocol}{minimum number of reads per protocol for parameter estimation}

\item{min_reads}{the minimum number of reads for Dirichlet parameter estimation}

\item{max_reads}{the maximum number of reads for Dirichlet parameter estimation}

\item{p_contamination}{the probability of a sample yielding significant contamination}

\item{show_plot}{show the clustering plot for sex chromosomal aneuploidy Dirichlet estimation?}

\item{printMissingIDs}{when combining karyotype calls, return names that are missing (or just the number of missing IDs)?}
}
\value{
summary
}
\description{
Title
}
