% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveChASM.R
\name{saveChASM}
\alias{saveChASM}
\title{A function for saving the results of a ChASM analysis as a tsv}
\usage{
saveChASM(inChASM, file, sort_by_samplename = FALSE)
}
\arguments{
\item{inChASM}{the result of a full ChASM analysis (output from runChASM)}

\item{file}{the full path and file name to place output}

\item{sort_by_samplename}{reorder alphabetically by sample names?}
}
\value{
saves results
}
\description{
A function for saving the results of a ChASM analysis as a tsv
}
\examples{
example_calls <- runChASM(rawReadCountsIn = example_data)
\dontrun{
saveChASM(inChASM = example_calls)
}
}
