% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostic.R
\name{plot_diagnostic}
\alias{plot_diagnostic}
\title{Plots diagnostic plots}
\usage{
plot_diagnostic(IDs, inChASM, addLabels = FALSE)
}
\arguments{
\item{IDs}{a list or predefined piped string of sample IDs to look for}

\item{inChASM}{parameter object}

\item{addLabels}{add "A", "B" and "C" to the plot panels?}
}
\value{
plots
}
\description{
Plots diagnostic plots
}
\examples{
example_calls <- runChASM(rawReadCountsIn = example_data)
plot_diagnostic(IDs = 'Ind_255_1', inChASM = example_calls, addLabels = TRUE)

}
